refactor main view
